testutil: move file content checker to separate file
