+++ testutil/filecontentchecker.go
